Split out edge, point, triangle classes into separate files Add <type> to collections
